start factor
